Moving integration tests
